move contractHost. Gallery needs to be moved into own package so that it can depend on both ERTP and Zoe.
