+++ packages/ERTP/test/shelved/peg/peg.js
@@ -2,7 +2,7 @@
 
 import harden from '@agoric/harden';
 import { makeNatDescOps } from '../../core/config/natDescOps';
-import { makeMint } from '../../core/mint';
+import { makeMint } from '../../../core/mint';
 
 // This implementation may be out of date
