+++ packages/ERTP/test/swingsetTests/presenceCorruption/vat-host.js
@@ -3,7 +3,7 @@
 import harden from '@agoric/harden';
 import evaluate from '@agoric/evaluate';
 
-import { makeContractHost } from '../../../core/contractHost';
+import { makeContractHost } from '../../shelved/contractHost/contractHost';
 
 function setup(syscall, state, helpers) {
   return helpers.makeLiveSlots(
